MTD-4-Refactored Rotation
Refactored Rotation code to GameActor
