+++ core/src/com/eric/mtd/game/model/actor/support/SupportActor.java
@@ -41,32 +41,6 @@ public class SupportActor extends GameActor implements Pool.Poolable, IAttacker{
 		this.gunPos = gunPos;
 		this.cost = cost;
 	}
-	public Vector2 getRotatedCoords(Vector2 coords) {
-		// Math stuff here -
-		// http://math.stackexchange.com/questions/270194/how-to-find-the-vertices-angle-after-rotation
-		double rotation = Math.toRadians(this.getRotation());
-		float cosa = (float) Math.cos(rotation);
-		float sina = (float) Math.sin(rotation);
-		float newX = ((((coords.x - getPositionCenter().x) * cosa) - ((coords.y - getPositionCenter().y) * sina)) + getPositionCenter().x);
-		float newY = ((((coords.x - getPositionCenter().x) * sina) + ((coords.y - getPositionCenter().y) * cosa)) + getPositionCenter().y);
-		return new Vector2(newX, newY);
-	}
-	public float calculateRotation(Vector2 vector) {
-		double prevAngle = this.getRotation();
-		double angle = MathUtils.atan2(getPositionCenter().x - vector.x, vector.y - getPositionCenter().y);
-		angle = Math.toDegrees(angle);
-		double negAngle = (angle - 360) % 360;
-		double posAngle = (angle + 360) % 360;
-		double negDistance = Math.abs(prevAngle - negAngle);
-		double posDistance = Math.abs(prevAngle - posAngle);
-		if (negDistance < posDistance) {
-			angle = negAngle;
-		} else {
-			angle = posAngle;
-		}
-		angle = Math.round(angle); // Round to help smooth movement
-		return (float) angle;
-	}
 	@Override
 	public void draw(Batch batch, float alpha) {
 		batch.end();
